requirements.txt
requests>=2.31,<3
pydantic>=2.5,<3
python-dotenv>=1.0,<2
rich>=13.7,<14
pytest>=8.0,<9
eval_type_backport>=0.2; python_version < "3.10"
